rest service-invoice-details + controller finished + bugs fixed

+++ src/main/java/com/api/eccom/repository/InvoiceRepository.java
@@ -9,5 +9,5 @@ import java.util.Optional;
 
 @Repository
 public interface InvoiceRepository extends JpaRepository<Invoice, Long> {
-    Optional<Invoice> findById(int id);
+    Optional<Invoice> findById(Long id);
 }
